Adding analyzers
